+++ src/main/java/com/nindybun/burnergun/common/items/upgrades/Trash/Trash.java
@@ -47,13 +47,13 @@ public class Trash extends UpgradeCard {
     public InteractionResultHolder<ItemStack> use(Level world, Player player, InteractionHand hand) {
         ItemStack stack = player.getItemInHand(hand);
         if (!world.isClientSide)
-            NetworkHooks.openGui((ServerPlayer) player, new SimpleMenuProvider(
+            /*NetworkHooks.openGui((ServerPlayer) player, new SimpleMenuProvider(
                     (windowId, playerInv, playerEntity) -> new TrashContainer(windowId, playerInv, (TrashHandler) getHandler(stack)),
-                    new TextComponent("")));
-            /*player.openMenu(new SimpleMenuProvider(
+                    new TextComponent("")));*/
+            player.openMenu(new SimpleMenuProvider(
                     (windowId, playerInv, playerEntity) -> new TrashContainer(windowId, playerInv, (TrashHandler) getHandler(stack)),
                     new TextComponent("")
-            ));*/
+            ));
         return InteractionResultHolder.success(stack);
     }