+++ src/main/java/com/nindybun/burnergun/common/network/PacketHandler.java
@@ -36,9 +36,8 @@ public class PacketHandler {
         INSTANCE.registerMessage(id++, PacketSpawnLightAtPlayer.class, PacketSpawnLightAtPlayer::encode, PacketSpawnLightAtPlayer::decode, PacketSpawnLightAtPlayer.Handler::handle);
         INSTANCE.registerMessage(id++, PacketSpawnLightAtRaycast.class, PacketSpawnLightAtRaycast::encode, PacketSpawnLightAtRaycast::decode, PacketSpawnLightAtRaycast.Handler::handle);
         INSTANCE.registerMessage(id++, PacketChangeColor.class, PacketChangeColor::encode, PacketChangeColor::decode, PacketChangeColor.Handler::handle);
-        INSTANCE.registerMessage(id++, PacketClientUpdateGun.class, PacketClientUpdateGun::encode, PacketClientUpdateGun::decode, PacketClientUpdateGun.Handler::handle);
+        INSTANCE.registerMessage(id++, PacketClientOpenGunSettings.class, PacketClientOpenGunSettings::encode, PacketClientOpenGunSettings::decode, PacketClientOpenGunSettings.Handler::handle);
         INSTANCE.registerMessage(id++, PacketClientPlayLightSound.class, PacketClientPlayLightSound::encode, PacketClientPlayLightSound::decode, PacketClientPlayLightSound.Handler::handle);
-        INSTANCE.registerMessage(id++, PacketClientRefuel.class, PacketClientRefuel::encode, PacketClientRefuel::decode, PacketClientRefuel.Handler::handle);
 
     }