+++ src/main/java/com/nindybun/burnergun/common/network/packets/PacketOpenAutoSmeltGui.java
@@ -53,23 +53,11 @@ public class PacketOpenAutoSmeltGui {
                 if (!(smelt.getItem() instanceof AutoSmelt))
                     return;
 
-                /*IItemHandler handler = smelt.getCapability(CapabilityItemHandler.ITEM_HANDLER_CAPABILITY).orElse(null);
+                IItemHandler handler = smelt.getCapability(CapabilityItemHandler.ITEM_HANDLER_CAPABILITY).orElse(null);
                 player.openMenu(new SimpleMenuProvider(
                         (windowId, playerInv, playerEntity) -> new AutoSmeltContainer(windowId, playerInv, (AutoSmeltHandler) handler),
                         new TextComponent("")
-                ));*/
-                ItemStackHandler ghostInventory = new ItemStackHandler(27) {
-                    @Override
-                    protected void onContentsChanged(int slot) {
-                        gun.getOrCreateTag().put(BurnerGunNBT.SMELTING_FILTER, serializeNBT());
-                    }
-                };
-
-                ghostInventory.deserializeNBT(gun.getOrCreateTagElement(BurnerGunNBT.SMELTING_FILTER));
-                player.openMenu(new SimpleMenuProvider(
-                        (windowId, playerInventory, playerEntity) -> new AutoSmeltContainer(windowId, playerInventory, ghostInventory), new TextComponent("")
                 ));
-
             });
 
             ctx.get().setPacketHandled(true);